+++ native_library/android/FirebaseANE/Crashlytics/src/main/java/com/tuarua/firebase/CrashlyticsANE.java
@@ -26,14 +26,10 @@ public class CrashlyticsANE implements FREExtension {
             ,"createGUID"
             ,"crash"
             ,"log"
-            ,"logException"
-            ,"setUserIdentifier"
-            ,"setUserEmail"
-            ,"setUserName"
-            ,"setString"
-            ,"setBool"
-            ,"setDouble"
-            ,"setInt"
+            ,"recordException"
+            ,"setUserId"
+            ,"setCustomKey"
+            ,"didCrashOnPreviousExecution"
     };
 
     private static CrashlyticsANEContext extensionContext;