+++ native_library/android/FirebaseANE/Vision/src/main/java/com/tuarua/firebase/VisionANE.java
@@ -25,11 +25,6 @@ public class VisionANE implements FREExtension {
     private static final String[] FUNCTIONS = {
              "init"
             ,"createGUID"
-            ,"requestPermissions"
-            ,"isCameraSupported"
-            ,"addNativeChild"
-            ,"updateNativeChild"
-            ,"removeNativeChild"
     };
     private static VisionANEContext extensionContext;