+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 0.13.0
+- AND: Analytics 17.4.4
+- AND: Crashlytics 17.1.1
+- AND: Firestore 21.5.0
+- AND: Remote Config 19.2.0
+- AND: Auth 19.3.2
+- AND: Play Services Auth 18.1.0
+- AND: Performance 19.0.8
+- AND: Messaging 20.2.3
+- AND: [Auth] - added `auth.signInWithCredential()`
+- AND: [Auth] - added `auth.signInWithProvider()`
+- AND: [Crashlytics] - now uses `FirebaseCrashlytics` API
+
 ### 0.12.0
 - Added OneSignal